+++ settings.gradle
@@ -1,3 +1,5 @@
 include ':FtcRobotController'
 include ':TeamCode'
 include ':MeepMeepTesting'
+include ':MeepMeepTesting'
+include ':MeepMeepTesting2'
